requirements.txt
python-telegram-bot==20.4
PyPDF2
python-docx
openai
